refactor(form-upload): add src and types folders
