+++ packages/form-upload/index.d.ts
@@ -1,9 +1,9 @@
-import Upload from './Upload';
-import FilePicker from './FilePicker';
-import FileList from './FileList';
-import FileRow from './FileRow';
-import FilePickerBtn from './FilePickerBtn';
-import UploadProgressBar from './UploadProgressBar';
+import Upload from './types/Upload';
+import FilePicker from './types/FilePicker';
+import FileList from './types/FileList';
+import FileRow from './types/FileRow';
+import FilePickerBtn from './types/FilePickerBtn';
+import UploadProgressBar from './types/UploadProgressBar';
 
 export default Upload;
