+++ packages/form-upload/package.json
@@ -4,6 +4,7 @@
   "author": "Tyson Warner <tyson.warner@availity.com>",
   "description": "Availity upload component for uploading files, compatible with @availity/form",
   "main": "index.js",
+  "types": "index.d.ts",
   "keywords": [
     "react",
     "availity",
@@ -40,6 +41,8 @@
     "uuid": "^8.3.2"
   },
   "devDependencies": {
+    "@availity/api-axios": ">=5.5.9",
+    "@availity/api-core": ">=6.6.4",
     "@availity/form": "^0.7.0",
     "@availity/upload-core": "^4.1.1",
     "formik": "^2.2.8",
